requirements.txt
asgiref==3.5.2
Django==4.0.6
django-crispy-forms==1.14.0
django-hCaptcha==0.2.0
django-utils-six==2.0
hurry.filesize==0.9
sqlparse==0.4.2
